+++ wabiz/client/wabiz/lib/core/core.dart
@@ -1,0 +1,2 @@
+export 'colors.dart';
+export 'theme.dart';
