+++ wabiz/client/wabiz/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:wabiz/core/theme.dart';
+import 'package:wabiz/core/core.dart';
 
 void main() {
   runApp(const MyApp());
@@ -12,10 +12,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
+      theme: wabizDefaultTheme,
       home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
